#130 refactor: SSE 연결 로직 수정
- useSSE 훅 제거, SSEProvider로 적용
